Implemented time function

+++ x64barebones/Kernel/include/lib.h
@@ -8,6 +8,6 @@ void * memcpy(void * destination, const void * source, uint64_t length);
 
 char *cpuVendor(char *result);
 
-uint64_t realTimeClock(int num);
+uint32_t realTimeClock(int num);
 
 #endif
